Facilitate merge method
Utilize Linux Kernel API to symplify the q_merge method and related
methods in queue.c.

There were two main defect in q_merge. First, there was instruction
directly accessing list_head structure defined in the API. However,
many of these was due to either my misconception such as pointing
list_head structure to null where it's already specified in the API,
or attempting to break the structure of doubly circular linked list.

After utilizing the API, there's no direct access to the structure
except the case of iterator.

Second, there was lots of if-else statement doing the same thing,
checking if a given queue is valid.

By considering not only simplicity of the code but also the correctness
of it, I replace sentinel controlled loop with counter controlled loop.
It prevent it from many if-else check.

Change-Id: I0b6d3b8ad1399fb9079955a92379ab109b0060c3